heap file draft done, scan.c not done
The problems I have so far:
1. Heapfile.c:  ~HeapFile()   how I know this file is a temp file?
2. Scan.h: the peekNext() func is wired.
3. do we need to define mvNext func?
4. how to unpin all pages in the constructor together?  traverse all
pages again?  wired.
5. Most Importantly: Whether we need to traverse the directory page to
scan, or sequential scan data pages only?
If the answer is no need, why there is a function called: nextDirPage()
If the answer is need dir, what is the meaning of scan. basically it
repeat operations in heap file.c

+++ HeapFile/src/heap_driver.C
@@ -88,7 +88,7 @@ int HeapDriver::test1()
         cout << "  - Add " << choice << " records to the file\n";
         for (int i =0; (i < choice) && (status == OK); i++)
           {
-            Rec rec = { i, i*2.5 };
+            Rec rec = { i, static_cast<float>(i*2.5) };  // original i*2.5    sixinglu 2016-02-19
             sprintf(rec.name, "record %i",i);
 
             status = f.insertRecord((char *)&rec, reclen, rid);
